M6 batch 1: STMG defaults, RTPC varint, runtime field order (M6-001/005/018)
M6-001: read the bank/HIRC nodes in the runtime's own field order.
- RTPC parameter id is a Wwise varint, not one byte; WwiseRtpc.ParamId widened to uint.
- A Sound source plug-in's u32 parameter size and bytes are consumed.
- Positioning (b0&b3), bus A/B/C conditional bodies, and the LayerCntr layer body
  fail explicitly rather than guessing an unrecovered length.
- The BKHD feedback flag (dword 3) is captured and adds four bytes to NodeBaseParams
  and to a bus.
- ReadBus rewritten to D6.1: no ranged bundle, A/B/C, maxInst, channel config,
  recovery, max duck, duck list, FX, mixer id, RTPCs, states.
- ReadBlend uses the real LayerCntr outer order.
- STMG defaults (M6-001): new WwiseStmg parses the Init.bnk state manager chunk
  recovered from the native reader at 0x9B0B14, exposing the RTPC default table
  (event_volume/robot_volume 1.0, 37 params).

M6-005: WwiseHash.Of bounded to the native 0x103-byte copy including the NUL.
M6-018: WwiseRuntimeSettings fixes the internal 48 kHz/1024 format (msPerFrame 21,
  +0xC 5, LPF chunk 128), distinct from the robot endpoint.

Tests: WwiseRuntimeTests pins the STMG layout, varint, source plug-in block, BKHD
flag, refused conditional bodies, LayerCntr order, hash bound and format constants.
WwiseTests.SoundDirs now falls back to the media directory so the shipped-bank
sweeps actually run instead of returning early; a new deterministic sweep loads the
six banks through WwiseAssets and asserts exact consumption. All 125 Wwise tests
pass.

+++ cozmo-stack/src/Cozmo.Robot/Animation/Wwise/WwiseSongRenderer.cs
@@ -433,7 +433,7 @@ public sealed class WwiseSongRenderer
         List<WwiseBoundModulator>? keep = null;
         foreach (var b in modulators)
         {
-            if (b.Binding.ParamId is (byte)WwiseProp.Volume or (byte)WwiseProp.Pitch) { (keep ??= new()).Add(b); continue; }
+            if (b.Binding.ParamId is (uint)WwiseProp.Volume or (uint)WwiseProp.Pitch) { (keep ??= new()).Add(b); continue; }
             problems.Add($"modulator {b.Modulator.Id} drives property {b.Binding.ParamId}, which the sampler does not apply");
         }
         return keep ?? NoModulators;
